Fix -Wsign-compare warnings: * introduce a local `index_type` alias where appropriate, to make index loop variables the same type as the type of the rhs of the loop comparison * static cast to std::size_t where appropriate in tests

+++ testsuite/allocatePopInternal.cpp
@@ -1,10 +1,10 @@
 /**
- * allocateSendPush.cpp - 
+ * allocateSendPush.cpp -
  * @author: Jonathan Beard
  * @version: Sat Feb 27 19:10:26 2016
- * 
+ *
  * Copyright 2016 Jonathan Beard
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at:
@@ -69,7 +69,7 @@ public:
     {
         obj_t in;
         input[ "x" ].pop( in );
-        assert( in == counter++ );
+        assert( static_cast<std::size_t>(in) == counter++ );
         return( raft::proceed );
     }
 
@@ -82,7 +82,7 @@ main( int argc, char **argv )
 {
     start s;
     last l;
-    
+
     raft::map M;
     M += s >> l;
     M.exe();
